Move helper method to somewhere central
